Main/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.44.0
django-bootstrap-form==3.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
